+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cap-config-writer
 
+## 0.7.29
+
+### Patch Changes
+
+-   Updated dependencies [34b2606]
+    -   @sap-ux/odata-service-inquirer@0.5.29
+
 ## 0.7.28
 
 ### Patch Changes